add middleware to the routes

+++ src/products/products.controller.ts
@@ -6,28 +6,46 @@ import {
   Put,
   Query,
   Delete,
+  Req,
+  UnauthorizedException,
 } from "@nestjs/common";
 import { ProductsService } from "./products.service";
 import { CreateProductDto } from "./dto/create-product.dto";
 import { UpdateProductDto } from "./dto/update-product.dto";
-import { ApiQuery } from "@nestjs/swagger";
+import {
+  ApiBearerAuth,
+  ApiOperation,
+  ApiQuery,
+  ApiResponse,
+} from "@nestjs/swagger";
 
 @Controller("products")
 export class ProductsController {
   constructor(private readonly productsService: ProductsService) {}
 
   @Post("create")
+  @ApiBearerAuth() // Add Bearer Auth to the Swagger docs
+  @ApiOperation({ summary: "Creates a product" })
+  @ApiResponse({ status: 200, description: "Product successfully removed" })
+  @ApiResponse({ status: 401, description: "Unauthorized access" })
   /**
    * Creates a new product.
    *
    * @param createProductDto - The create product DTO.
    * @returns The created product.
    */
-  create(@Body() createProductDto: CreateProductDto) {
+  create(@Body() createProductDto: CreateProductDto, @Req() req: any) {
+    if (req.userRole !== "admin") {
+      throw new UnauthorizedException("Only admins can access this route");
+    }
     return this.productsService.create(createProductDto);
   }
 
   @Get("get")
+  @ApiOperation({ summary: "Updates a product" })
+  @ApiResponse({ status: 200, description: "Product found" })
+  @ApiResponse({ status: 401, description: "Unauthorized access" })
+  @ApiResponse({ status: 404, description: "Product not found" })
   /**
    * Gets a product by product code or location.
    *
@@ -45,6 +63,11 @@ export class ProductsController {
   }
 
   @Put("update")
+  @ApiBearerAuth() // Add Bearer Auth to the Swagger docs
+  @ApiOperation({ summary: "Updates a product" })
+  @ApiResponse({ status: 200, description: "Product successfully updated" })
+  @ApiResponse({ status: 401, description: "Unauthorized access" })
+  @ApiResponse({ status: 404, description: "Product not found" })
   /**
    * Updates a product.
    *
@@ -56,12 +79,21 @@ export class ProductsController {
    */
   update(
     @Query("productCode") productCode: number,
-    @Body() updateProductDto: UpdateProductDto
+    @Body() updateProductDto: UpdateProductDto,
+    @Req() req: any
   ) {
+    if (req.userRole !== "admin") {
+      throw new UnauthorizedException("Only admins can access this route");
+    }
     return this.productsService.update(productCode, updateProductDto);
   }
 
   @Delete("remove")
+  @ApiBearerAuth() // Add Bearer Auth to the Swagger docs
+  @ApiOperation({ summary: "Removes a product" })
+  @ApiResponse({ status: 200, description: "Product successfully removed" })
+  @ApiResponse({ status: 401, description: "Unauthorized access" })
+  @ApiResponse({ status: 404, description: "Product not found" })
   /**
    * Removes a product.
    *
@@ -70,7 +102,10 @@ export class ProductsController {
    * @throws {NotFoundException} If no product code is provided,
    *                              or if the product is not found.
    */
-  remove(@Query("productCode") productCode: number) {
+  remove(@Query("productCode") productCode: number, @Req() req: any) {
+    if (req.userRole !== "admin") {
+      throw new UnauthorizedException("Only admins can access this route");
+    }
     return this.productsService.remove(productCode);
   }
 }
